requirements.txt
accelerate==0.34.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
bcrypt==5.0.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.6
captcha==0.7.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
datasets==2.21.0
decorator==5.2.1
diffusers==0.30.0
dill==0.3.8
distro==1.9.0
extra-streamlit-components==0.1.81
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-genai==1.60.0
gradio==6.4.0
gradio_client==2.0.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
inflate64==1.0.3
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.16
multivolumefile==0.2.3
narwhals==2.16.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.1.105
orjson==3.11.5
packaging==26.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.1
plotly==6.5.2
pooch==1.8.2
proglog==0.1.12
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
py7zr==1.0.0
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybcj==1.0.6
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.11.0
pyngrok==7.5.0
pyppmd==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzstd==0.18.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
sentry-sdk==2.50.0
setuptools==80.10.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.50.0
streamlit==1.54.0
streamlit-authenticator==0.4.2
sympy==1.14.0
tangoflux @ git+https://github.com/declare-lab/TangoFlux@fb364c254884c017d327ff5180660ec59ba56251
tenacity==9.1.2
texttable==1.7.0
threadpoolctl==3.6.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.13.3
torch==2.4.0
torchaudio==2.4.0
torchlibrosa==0.1.0
torchvision==0.19.0
tornado==6.5.4
tqdm==4.67.1
transformers==4.44.0
triton==3.0.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wandb==0.24.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.46.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
